requirements.txt
flask
youtube_transcript_api
google-api-python-client
google-generativeai
googletrans==3.1.0a0
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2